+++ src/shared/Router.js
@@ -10,12 +10,11 @@ import {
   PetDetail,
   Pets,
   SurveyForm,
+  SurveyDone,
   Signup,
   Main
 } from '../pages';
-import SurveyDone from '../pages/surveyForm/SurveyDone';
 import { GlobalLayout } from '../common';
-import Petdetailcopy from '../pages/petDetail/Petdetail copy';
 
 const Router = () => {
   return (
@@ -30,10 +29,8 @@ const Router = () => {
       <Route path="/mypage" element={<Mypage />} />
       <Route path="/mypage/edit" element={<EditMyapge />} />
       <Route path="/pets" element={<Pets />} />
-      <Route path="/pets/:id" element={<PetDetail />} />
       <Route path="/surveyform" element={<SurveyForm />} />
       <Route path="/surveyDone" element={<SurveyDone />} />
-      <Route path="/petsdetailcopy" element={<Petdetailcopy />} />
 
       <Route path="/login" element={<Login />} />
       <Route path="/signup" element={<Signup />} />
